chore: Reorganize passwordReq and sidebar components
