+++ src/app/owner/layout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Sidebar from "~components/sidebar";
+import Sidebar from "~components/sidebar/sidebar";
 import styles from "../layout.module.scss";
 
 export default function OwnerLayout({
